+++ release_notes.md
@@ -377,7 +377,8 @@ This release implements performance bug fixes and provides Data object cache fil
 # v1.4.0
 
 ## Feature
-1. Added `clear_cache` method to data object to remove cache file. 
+1. [TH2-4601] Added `clear_cache` method to data object to remove cache file.
 
 ## BugFixes
-1. [TH2-4604] `Data.from_cache_file` method has return value now.
+1. [TH2-4603] Fix += joining feature. Now it keeps cache status.
+2. [TH2-4604] `Data.from_cache_file` method has return value now.